+++ src/components/mypage/bookmark/favoriteProject/FavoriteProjectList.tsx
@@ -1,6 +1,6 @@
 import styles from './favoriteProjectCard.module.css'
 import FavoriteProjectCard from './FavoriteProjectCard'
-import { FavoriteProjectListProps } from '@/types/favoriteType'
+import { FavoriteProjectListProps } from '@/types/mypageDataType'
 import FavoriteNull from '../favoriteNull/FavoriteNull'
 
 const FavoriteProjectList = () => {
